Move debug util to separate pkg to avoid import loop
